requirements.txt
reflex>=0.6.0
h2ogpte>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
